commit 修改版本

+++ app/src/main/java/com/jieniuwuliu/jieniu/bean/LoginBean.java
@@ -366,8 +366,8 @@ public class LoginBean {
             private int uid;
             private String name;
             private String phone;
-            private int lng;
-            private int lat;
+            private double lng;
+            private double lat;
 
             public int getId() {
                 return id;
@@ -433,18 +433,18 @@ public class LoginBean {
                 this.phone = phone;
             }
 
-            public int getLng() {
+            public double getLng() {
                 return lng;
             }
 
-            public void setLng(int lng) {
+            public void setLng(double lng) {
                 this.lng = lng;
             }
 
-            public int getLat() {
+            public double getLat() {
                 return lat;
             }
-            public void setLat(int lat) {
+            public void setLat(double lat) {
                 this.lat = lat;
             }
